splitting clap related code on apps in bin/
