+++ src/cli/mod.rs
@@ -1,8 +1,3 @@
 pub mod cli_config;
-pub mod cli_parser;
 
 pub use cli_config::{CashierdConfig, Config, DarkfidConfig, DrkConfig, GatewaydConfig, MapConfig};
-
-pub use cli_parser::{
-    CliCashierd, CliDao, CliDaoSubCommands, CliDarkfid, CliDrk, CliDrkSubCommands, CliGatewayd,
-};
